[ticket/13961] Move service_collection to di/service_collection namespace
PHPBB3-13961
