+++ phpBB/config/default/container/services_console.yml
@@ -1,6 +1,6 @@
 services:
     console.command_collection:
-        class: phpbb\di\service_collection
+        class: phpbb\di\service_collection\service_collection
         arguments:
             - @service_container
         tags:
