+++ phpBB/config/default/container/services_mimetype_guesser.yml
@@ -1,6 +1,6 @@
 services:
     mimetype.guesser_collection:
-        class: phpbb\di\service_collection
+        class: phpbb\di\service_collection\service_collection
         arguments:
             - @service_container
         tags:
