+++ phpBB/config/default/container/services_password.yml
@@ -18,7 +18,7 @@ services:
 
 # ----- Password's drivers -----
     passwords.driver_collection:
-        class: phpbb\di\service_collection
+        class: phpbb\di\service_collection\service_collection
         arguments:
             - @service_container
         tags:
