+++ phpBB/config/default/container/services_profilefield.yml
@@ -21,7 +21,7 @@ services:
 
 # ----- Profile fields types -----
     profilefields.type_collection:
-        class: phpbb\di\service_collection
+        class: phpbb\di\service_collection\service_collection
         arguments:
             - @service_container
         tags:
