+++ phpBB/phpbb/captcha/factory.php
@@ -21,7 +21,7 @@ class factory
 	private $container;
 
 	/**
-	* @var \phpbb\di\service_collection
+	* @var \phpbb\di\service_collection\service_collection
 	*/
 	private $plugins;
 
@@ -29,9 +29,9 @@ class factory
 	* Constructor
 	*
 	* @param \Symfony\Component\DependencyInjection\ContainerInterface $container
-	* @param \phpbb\di\service_collection                              $plugins
+	* @param \phpbb\di\service_collection\service_collection                              $plugins
 	*/
-	public function __construct(\Symfony\Component\DependencyInjection\ContainerInterface $container, \phpbb\di\service_collection $plugins)
+	public function __construct(\Symfony\Component\DependencyInjection\ContainerInterface $container, \phpbb\di\service_collection\service_collection $plugins)
 	{
 		$this->container = $container;
 		$this->plugins = $plugins;
